+++ crates/astria-sequencer/src/action_handler/impls/bridge_lock.rs
@@ -1,0 +1,97 @@
+use astria_core::{
+    protocol::transaction::v1::action::{
+        BridgeLock,
+        Transfer,
+    },
+    sequencerblock::v1::block::Deposit,
+};
+use astria_eyre::eyre::{
+    ensure,
+    OptionExt as _,
+    Result,
+    WrapErr as _,
+};
+use async_trait::async_trait;
+use cnidarium::StateWrite;
+
+use crate::{
+    action_handler::{
+        check_transfer,
+        execute_transfer,
+        ActionHandler,
+    },
+    address::StateReadExt as _,
+    bridge::{
+        StateReadExt as _,
+        StateWriteExt,
+    },
+    transaction::StateReadExt as _,
+    utils::create_deposit_event,
+};
+
+#[async_trait]
+impl ActionHandler for BridgeLock {
+    async fn check_stateless(&self) -> Result<()> {
+        Ok(())
+    }
+
+    async fn check_and_execute<S: StateWrite>(&self, mut state: S) -> Result<()> {
+        let from = state
+            .get_transaction_context()
+            .expect("transaction source must be present in state when executing an action")
+            .address_bytes();
+        state
+            .ensure_base_prefix(&self.to)
+            .await
+            .wrap_err("failed check for base prefix of destination address")?;
+        // ensure the recipient is a bridge account.
+        let rollup_id = state
+            .get_bridge_account_rollup_id(&self.to)
+            .await
+            .wrap_err("failed to get bridge account rollup id")?
+            .ok_or_eyre("bridge lock must be sent to a bridge account")?;
+
+        let allowed_asset = state
+            .get_bridge_account_ibc_asset(&self.to)
+            .await
+            .wrap_err("failed to get bridge account asset ID")?;
+        ensure!(
+            allowed_asset == self.asset.to_ibc_prefixed(),
+            "asset ID is not authorized for transfer to bridge account",
+        );
+
+        let source_transaction_id = state
+            .get_transaction_context()
+            .expect("current source should be set before executing action")
+            .transaction_id;
+        let source_action_index = state
+            .get_transaction_context()
+            .expect("current source should be set before executing action")
+            .source_action_index;
+
+        let deposit = Deposit {
+            bridge_address: self.to,
+            rollup_id,
+            amount: self.amount,
+            asset: self.asset.clone(),
+            destination_chain_address: self.destination_chain_address.clone(),
+            source_transaction_id,
+            source_action_index,
+        };
+        let deposit_abci_event = create_deposit_event(&deposit);
+
+        let transfer_action = Transfer {
+            to: self.to,
+            asset: self.asset.clone(),
+            amount: self.amount,
+            fee_asset: self.fee_asset.clone(),
+        };
+
+        check_transfer(&transfer_action, &from, &state).await?;
+        execute_transfer(&transfer_action, &from, &mut state).await?;
+
+        state.cache_deposit_event(deposit);
+        state.record(deposit_abci_event);
+        Ok(())
+    }
+}
