+++ crates/astria-sequencer/src/action_handler/impls/fee_asset_change.rs
@@ -1,0 +1,65 @@
+use astria_core::protocol::transaction::v1::action::FeeAssetChange;
+use astria_eyre::eyre::{
+    self,
+    ensure,
+    WrapErr as _,
+};
+use async_trait::async_trait;
+use cnidarium::StateWrite;
+use futures::StreamExt as _;
+use tokio::pin;
+
+use crate::{
+    action_handler::ActionHandler,
+    authority::StateReadExt as _,
+    fees::{
+        StateReadExt as _,
+        StateWriteExt as _,
+    },
+    transaction::StateReadExt as _,
+};
+
+#[async_trait]
+impl ActionHandler for FeeAssetChange {
+    async fn check_stateless(&self) -> eyre::Result<()> {
+        Ok(())
+    }
+
+    async fn check_and_execute<S: StateWrite>(&self, mut state: S) -> eyre::Result<()> {
+        let from = state
+            .get_transaction_context()
+            .expect("transaction source must be present in state when executing an action")
+            .address_bytes();
+        let authority_sudo_address = state
+            .get_sudo_address()
+            .await
+            .wrap_err("failed to get authority sudo address")?;
+        ensure!(
+            authority_sudo_address == from,
+            "unauthorized address for fee asset change"
+        );
+        match self {
+            FeeAssetChange::Addition(asset) => {
+                state
+                    .put_allowed_fee_asset(asset)
+                    .context("failed to write allowed fee asset to state")?;
+            }
+            FeeAssetChange::Removal(asset) => {
+                state.delete_allowed_fee_asset(asset);
+
+                pin!(
+                    let assets = state.allowed_fee_assets();
+                );
+                ensure!(
+                    assets
+                        .filter_map(|item| std::future::ready(item.ok()))
+                        .next()
+                        .await
+                        .is_some(),
+                    "cannot remove last allowed fee asset",
+                );
+            }
+        }
+        Ok(())
+    }
+}
