+++ crates/astria-sequencer/src/action_handler/impls/fee_change.rs
@@ -1,0 +1,82 @@
+use astria_core::protocol::transaction::v1::action::FeeChange;
+use astria_eyre::eyre::{
+    self,
+    ensure,
+    WrapErr as _,
+};
+use async_trait::async_trait;
+use cnidarium::StateWrite;
+
+use crate::{
+    action_handler::ActionHandler,
+    authority::StateReadExt as _,
+    fees::StateWriteExt as _,
+    transaction::StateReadExt as _,
+};
+
+#[async_trait]
+impl ActionHandler for FeeChange {
+    async fn check_stateless(&self) -> eyre::Result<()> {
+        Ok(())
+    }
+
+    /// check that the signer of the transaction is the current sudo address,
+    /// as only that address can change the fee
+    async fn check_and_execute<S: StateWrite>(&self, mut state: S) -> eyre::Result<()> {
+        let from = state
+            .get_transaction_context()
+            .expect("transaction source must be present in state when executing an action")
+            .address_bytes();
+        // ensure signer is the valid `sudo` key in state
+        let sudo_address = state
+            .get_sudo_address()
+            .await
+            .wrap_err("failed to get sudo address from state")?;
+        ensure!(sudo_address == from, "signer is not the sudo key");
+
+        match self {
+            Self::Transfer(fees) => state
+                .put_transfer_fees(*fees)
+                .wrap_err("failed to put transfer fees"),
+            Self::RollupDataSubmission(fees) => state
+                .put_rollup_data_submission_fees(*fees)
+                .wrap_err("failed to put sequence fees"),
+            Self::Ics20Withdrawal(fees) => state
+                .put_ics20_withdrawal_fees(*fees)
+                .wrap_err("failed to put ics20 withdrawal fees"),
+            Self::InitBridgeAccount(fees) => state
+                .put_init_bridge_account_fees(*fees)
+                .wrap_err("failed to put init bridge account fees"),
+            Self::BridgeLock(fees) => state
+                .put_bridge_lock_fees(*fees)
+                .wrap_err("failed to put bridge lock fees"),
+            Self::BridgeUnlock(fees) => state
+                .put_bridge_unlock_fees(*fees)
+                .wrap_err("failed to put bridge unlock fees"),
+            Self::BridgeSudoChange(fees) => state
+                .put_bridge_sudo_change_fees(*fees)
+                .wrap_err("failed to put bridge sudo change fees"),
+            Self::IbcRelay(fees) => state
+                .put_ibc_relay_fees(*fees)
+                .wrap_err("failed to put ibc relay fees"),
+            Self::ValidatorUpdate(fees) => state
+                .put_validator_update_fees(*fees)
+                .wrap_err("failed to put validator update fees"),
+            Self::FeeAssetChange(fees) => state
+                .put_fee_asset_change_fees(*fees)
+                .wrap_err("failed to put fee asset change fees"),
+            Self::FeeChange(fees) => state
+                .put_fee_change_fees(*fees)
+                .wrap_err("failed to put fee change fees"),
+            Self::IbcRelayerChange(fees) => state
+                .put_ibc_relayer_change_fees(*fees)
+                .wrap_err("failed to put ibc relayer change fees"),
+            Self::SudoAddressChange(fees) => state
+                .put_sudo_address_change_fees(*fees)
+                .wrap_err("failed to put sudo address change fees"),
+            Self::IbcSudoChange(fees) => state
+                .put_ibc_sudo_change_fees(*fees)
+                .wrap_err("failed to put ibc sudo change fees"),
+        }
+    }
+}
